Corrected data types for year and tacts count; changed indication for set second mode. All modules are working correctly.

+++ ClockController.cpp
@@ -85,6 +85,6 @@ class ClockController
   uint8_t _second;
   uint8_t _day;
   uint8_t _month;
-  uint8_t _year;
+  uint16_t _year;
   uint8_t _weekday;
 };
